We assume that our floats fit into one of our instruction params (which is a 32-bit int). So we need to use floats, not doubles, otherwise we'll hit the assert that enforces this as soon as we use a floating point constant in code. To do: Maybe make instruction params a union of float and int. Will make bytecode use more RAM, but it's also less of a low-level hack and less restrictive.

+++ common/LEOValue.h
@@ -77,7 +77,7 @@ typedef unsigned long			LEOObjectSeed;
 typedef long long				LEOInteger;
 
 /*! The type of fractional numbers in the language. */
-typedef double					LEONumber;
+typedef float					LEONumber;
 
 
 /*! A unit attached to a numerical value. */
